% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multipass.cppls.R
\name{multipass.cppls}
\alias{multipass.cppls}
\title{Wrapper for cppls function with multiple pass and outliers elmination}
\usage{
multipass.cppls(formula, ncomp, validation, data, nbpass, out = 2.5, Y.add)
}
\arguments{
\item{formula}{a model formula. Most of the lm formula constructs are supported. See below.}

\item{ncomp}{the number of components to include in the model (see below).}

\item{validation}{character. What kind of (internal) validation to use}

\item{data}{an optional data frame with the data to fit the model from.}

\item{nbpass}{Number of passes}

\item{out}{T value threshold for outlier elimination}

\item{Y.add}{a vector or matrix of additional responses containing relevant information about the observations. Only used for cppls.}
}
\value{
A list with the following components

outliers A vector with outliers names as deduced from data row.names

plsr the pls model object from the final pass

ncomp the number of components to be further used

pass a vector of same length as outliers indicating at which pass outliers were removed
}
\description{
Wrapper for cppls function with multiple pass and outliers elmination
}
\examples{
\dontrun{
}
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

