% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multipass.pls.R
\name{multipass.pls}
\alias{multipass.pls}
\title{Wrapper for plsr function with multiple pass and outliers elmination}
\usage{
multipass.pls(formula, ncomp, validation, data, nbpass, out = 2.5)
}
\arguments{
\item{formula}{a model formula. Most of the lm formula constructs are supported. See below.}

\item{ncomp}{the number of components to include in the model (see below).}

\item{validation}{character. What kind of (internal) validation to use}

\item{data}{an optional data frame with the data to fit the model from.}

\item{nbpass}{Number of passes}

\item{out}{T value threshold for outlier elimination}
}
\value{
A list with the following components

outliers A vector with outliers names as deduced from data row.names

plsr the pls model object from the final pass

ncomp the number of components to be further used

pass a vector of same length as outliers indicating at which pass outliers were removed
}
\description{
Wrapper for plsr function with multiple pass and outliers elmination
}
\examples{
\dontrun{
}
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

